baselines/WEB-013/requirements.txt
# GraphQL API with N+1 Prevention Requirements

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.1

# GraphQL
graphql-core==3.2.3
strawberry-graphql[fastapi]==0.217.1
graphene==3.3

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
databases==0.8.0

# DataLoader
aiodataloader==0.4.0

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# Redis (optional caching)
redis==5.0.1
hiredis==2.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
pylint==3.0.3
mypy==1.8.0